+++ grails-app/conf/BuildConfig.groovy
@@ -10,6 +10,7 @@ grails.project.class.dir = "target/classes"
 grails.project.test.class.dir = "target/test-classes"
 grails.project.test.reports.dir = "target/test-reports"
 
+grails.project.dependency.resolver = "maven"
 grails.project.dependency.resolution = {
     inherits("global") {
     }
@@ -23,11 +24,13 @@ grails.project.dependency.resolution = {
     }
 
     plugins {
-        compile ":taggable:1.0.1"
-        compile (":spring-security-core:1.2.7.3") {
+        compile (":spring-security-core:1.2.7.3", ":taggable:1.0.1") {
             export = false
         }
-        build (":tomcat:$grailsVersion", ":release:2.0.3", ":rest-client-builder:1.0.2", ":resources:1.2.RC2") {
+        runtime (":resources:1.2.1") {
+            export = false
+        }
+        build (":tomcat:$grailsVersion", ":release:2.0.3", ":rest-client-builder:1.0.2") {
             export = false
         }
     }